requirements.txt
wheel>=0.38.0
pandas
openai
pydantic
psycopg
python-dotenv
timescale-vector
instructor
anthropic
fastapi
uvicorn
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
git+https://github.com/openai/whisper.git
pyaudio
playsound==1.2.2
tts==0.22.0
pytest
python-multipart
-e .
dateparser